requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==3.6.4.0
celery==5.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.14
Django==4.0.1
django-allauth==0.63.1
django-cors-headers==3.10.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
idna==3.7
kombu==5.3.7
packaging==24.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
pytz==2024.1
redis==4.1.0
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
stripe==2.65.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
